+++ src/abagames/gr/soundmanager.d
@@ -31,8 +31,8 @@ public class SoundManager: abagames.util.sdl.sound.SoundManager {
   Rand rand;
   string[] bgmFileName;
   string currentBgm;
-  int prevBgmIdx;
-  int nextIdxMv;
+  size_t prevBgmIdx;
+  size_t nextIdxMv;
 
   public static void setRandSeed(long seed) {
     rand.setSeed(seed);
@@ -81,14 +81,14 @@ public class SoundManager: abagames.util.sdl.sound.SoundManager {
   }
 
   public static void playBgm() {
-    int bgmIdx = rand.nextInt(bgm.length - RANDOM_BGM_START_INDEX) + RANDOM_BGM_START_INDEX;
+    size_t bgmIdx = rand.nextInt(bgm.length - RANDOM_BGM_START_INDEX) + RANDOM_BGM_START_INDEX;
     nextIdxMv = rand.nextInt(2) * 2 - 1;
     prevBgmIdx = bgmIdx;
     playBgm(bgmFileName[bgmIdx]);
   }
 
   public static void nextBgm() {
-    int bgmIdx = prevBgmIdx + nextIdxMv;
+    size_t bgmIdx = prevBgmIdx + nextIdxMv;
     if (bgmIdx < RANDOM_BGM_START_INDEX)
       bgmIdx = bgm.length - 1;
     else if (bgmIdx >= bgm.length)
